chore: cleanup syntax
