+++ src/features/search/components/SearchResult.tsx
@@ -1,26 +1,6 @@
-import { JSX } from "react";
-
-import { DocumentText } from "@utils/interface";
-import { useFetchQueryResult } from "../api/getQueryResult";
 import Spinner from "@/components/Spinner";
-
-export function parseHighlightText(documentText: DocumentText) {
-  const text = documentText.Text;
-  const parsedText: (string | JSX.Element)[] = [];
-
-  let cur = 0;
-  documentText.Highlights.forEach((highlight) => {
-    parsedText.push(text.substring(cur, highlight.BeginOffset));
-    parsedText.push(
-      <b>{text.substring(highlight.BeginOffset, highlight.EndOffset)}</b>
-    );
-    cur = highlight.EndOffset;
-  });
-
-  if (cur < text.length) parsedText.push(text.substring(cur));
-
-  return <p>{...parsedText}</p>;
-}
+import { useFetchQueryResult } from "../api/getQueryResult";
+import { parseHighlightText } from "../utils/parseHighlight";
 
 type PropType = {
   searchString: string;
